% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.v1.R
\name{coverage_rate}
\alias{coverage_rate}
\title{Coverage Rate Calculation}
\usage{
coverage_rate(mod.o, mod.f)
}
\arguments{
\item{mod.o}{A list containing different reference sets.}

\item{mod.f}{A list containing different test sets.}
}
\value{
a numeric as the overall coverage rate.
}
\description{
Calculates coverage rate.
}
\examples{
mod.o = list(set1 = LETTERS[1:10])
mod.f = list(set1 = LETTERS[5:20])
coverage_rate(mod.o,mod.f)
}
