% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.v1.R
\name{detection_accuracy}
\alias{detection_accuracy}
\title{Detection accuracy Calculation}
\usage{
detection_accuracy(mod.o, mod.f)
}
\arguments{
\item{mod.o}{A list containing different reference sets.}

\item{mod.f}{A list containing different test sets.}
}
\value{
a numeric as the overall detection accuracy.
}
\description{
Calculates overall detection accuracy.
}
\examples{
mod.o = list(set1 = LETTERS[1:10])
mod.f = list(set1 = LETTERS[5:20])
detection_accuracy(mod.o,mod.f)
}
