% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dropout_rates.v1.R
\name{get_dropout_rates}
\alias{get_dropout_rates}
\title{Dropout rate inference}
\usage{
get_dropout_rates(mat, gene.names)
}
\arguments{
\item{mat}{single-cell count matrix log-normalized gene expression matrix to calculate the average expression levels.}

\item{gene.names}{rownames of the count matrix to run the dropout inference.}
}
\value{
a matrix of inferred dropout rate per gene per cell.
}
\description{
Uses scImpute framework to estimate dropout rates per cell, per gene using Gamma-normal mixture model.
}
\examples{
# See Vignettes. 
}
