% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.v1.R
\name{inclusion_rate}
\alias{inclusion_rate}
\title{Inclusion Rate Calculation}
\usage{
inclusion_rate(mod.o, mod.f)
}
\arguments{
\item{mod.o}{A list containing different reference sets.}

\item{mod.f}{A list containing different test sets.}
}
\value{
a numeric as the overall inclusion rate.
}
\description{
Calculates inclusion rate.
}
\examples{
mod.o = list(set1 = LETTERS[1:10])
mod.f = list(set1 = LETTERS[5:20])
inclusion_rate(mod.o,mod.f)
}
