% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_8k_msc_results}
\alias{pbmc_8k_msc_results}
\title{MSC results from 8k PBMC data from 10x benchmark data set}
\format{
\subsection{\code{pbmc_8k_msc_results}}{

A list containing the MSC results
\describe{
\item{modules}{A list of cell clusters}
\item{module.table}{A data.frame containing cell cluster statistics (sizes, compactness, intra-cluster connectivity etc).}
\item{cell.network}{An igraph object containing the cell network}
\item{alpha.value}{The alpha parameter value used to calculate the cluster compactness}
\item{pruned.table}{Refined list of cell clusters after applying compactness and intra-cluster connectivity filters.}
}
}
}
\source{
\url{https://www.10xgenomics.com/resources/datasets/8-k-pbm-cs-from-a-healthy-donor-2-standard-2-0-1}
}
\usage{
pbmc_8k_msc_results
}
\description{
MSC results from 8k PBMC data from 10x benchmark data set
}
\keyword{datasets}
