% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.v3.4.R,
%   R/clustering_functions.v3.5.R
\docType{package}
\name{prune_hierarchy.compact}
\alias{prune_hierarchy.compact}
\title{Pruning cluster hierarchy}
\usage{
prune_hierarchy.compact(iter.res)

prune_hierarchy.compact(iter.res)
}
\arguments{
\item{iter.res}{An output from iterative_clustering or iterative_clustering.par.}
}
\value{
Adds "prunted.table" into the input data.

Adds "prunted.table" into the input data.
}
\description{
Quick filtering of module hierarchy tree to remove branches that do not anchor on the grand parent cluster, M0.

Quick filtering of module hierarchy tree to remove branches that do not anchor on the grand parent cluster, M0.
}
\seealso{
\code{\link[=iterative_clustering]{iterative_clustering()}} or \code{\link[=iterative_clustering.par]{iterative_clustering.par()}} for the input.

\code{\link[=iterative_clustering]{iterative_clustering()}} or \code{\link[=iterative_clustering.par]{iterative_clustering.par()}} for the input.
}
