% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simMix1}
\alias{simMix1}
\title{simMix1 data set from pipeComp study}
\format{
\subsection{\code{simMix1}}{

A list containing the MSC results
\describe{
\item{simMix}{SingleCellExperiment object}
}
}
}
\source{
\url{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-020-02136-7}
}
\usage{
simMix1
}
\description{
simMix1 data set from pipeComp study
}
\keyword{datasets}
